Changed file structure, class paths, and name of our user collection to userCourses. Works.
