+++ imports/ui/components/dashboard/AddCourse.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Container, Form, Button, Header, Popup, Grid } from 'semantic-ui-react'
 import { createContainer } from "meteor/react-meteor-data";
-import { Users } from "../../../collections/users";
+import { Users } from "../../../collections/userCourses";
 const shortid = require("shortid");
 
 class AddCourse extends React.Component {
