+++ imports/ui/components/dashboard/ProgressViews/Progress2.js
