+++ imports/ui/components/dashboard/ProgressViews/ProgressTabs.js
